+++ ydb/core/persqueue/pqtablet/partition/user_info.h
@@ -3,7 +3,7 @@
 #include "account_read_quoter.h"
 
 #include <ydb/core/persqueue/working_time_counter.h>
-#include <ydb/core/persqueue/subscriber.h>
+#include "subscriber.h"
 #include <ydb/core/persqueue/percentile_counter.h>
 #include "quota_tracker.h"
 #include <ydb/core/persqueue/public/constants.h>
